Update 2019-02-07-programming-camp.md
Are we set on incorporating calculus applications into this lecture as well? If so, let's make some adjustments to the title/description to reflect this. Like you mentioned during today's meeting, not only is it easy enough to cover both in a 2-hour span, I think it would be valuable to knock both out in a timely manner like this.